+++ packages/Helm/src/package-test.ts
@@ -1,10 +1,13 @@
 import * as DG from 'datagrok-api/dg';
 import * as grok from 'datagrok-api/grok';
 
-import '@datagrok-libraries/bio/src/types/helm';
+// Do not import anything from JsDrawLite or HelmWebEditor, only to the main Helm package
+// import {JSDraw2ModuleType} from '@datagrok/js-draw-lite/src/types/jsdraw2';
+// import {HelmType, OrgHelmModuleType} from '@datagrok/helm-web-editor/src/types/org-helm';
+
 import {runTests, tests, TestContext} from '@datagrok-libraries/utils/src/test';
 
-import './tests/helm-tests.ts';
+import './tests/helm-tests';
 import './tests/findMonomers-tests';
 import './tests/helm-service-tests';
 import './tests/renderers-tests';
@@ -12,6 +15,8 @@ import './tests/get-molfiles-tests';
 import './tests/properties-widget-tests';
 import './tests/get-all-parts-tests';
 import './tests/get-monomer-tests';
+import './tests/parse-helm-tests';
+import './tests/helm-web-editor-tests';
 
 export const _package = new DG.Package();
 export {tests};