+++ packages/Helm/src/package-utils.ts
@@ -2,19 +2,27 @@ import * as ui from 'datagrok-api/ui';
 import * as grok from 'datagrok-api/grok';
 import * as DG from 'datagrok-api/dg';
 
-import * as org from 'org';
-import * as scil from 'scil';
-import * as JSDraw2 from 'JSDraw2';
-import Atom = JSDraw2.Atom;
-
+import {delay, timeout} from '@datagrok-libraries/utils/src/test';
 import {errInfo} from '@datagrok-libraries/bio/src/utils/err-info';
+import type {
+  DojoType, DojoxType, IOrgHelmWebEditor, Monomers
+} from '@datagrok-libraries/bio/src/helm/types';
 import {HelmServiceBase} from '@datagrok-libraries/bio/src/viewers/helm-service';
-import {IMonomerLib, HelmType} from '@datagrok-libraries/bio/src/types';
-import {ILogger} from '@datagrok-libraries/bio/src/utils/logger';
+import {IMonomerLib} from '@datagrok-libraries/bio/src/types';
 
+import {HelmHelper} from './helm-helper';
 import {HelmService} from './utils/helm-service';
-import {GetMonomerFunc, GetMonomerResType, getWebEditorMonomer} from './utils/get-monomer';
+import {OrgHelmModule, ScilModule} from './types';
 
+import {_package} from './package';
+
+declare const dojo: DojoType;
+declare const dojox: DojoxType;
+declare const scil: ScilModule;
+declare const org: OrgHelmModule;
+
+// eslint-disable-next-line
+var dojoConfig = {async: true};
 
 type HelmWindowType = Window & {
   $helmService?: HelmServiceBase,
@@ -27,11 +35,48 @@ export function _getHelmService(): HelmServiceBase {
   return res;
 }
 
-export function initHelmPatchDojo(): void {
+export async function initHelmLoadAndPatchDojo(): Promise<void> {
+  const logPrefix = `Helm: _package.initHelmPatchDojo()`;
   // patch window.dojox.gfx.svg.Text.prototype.getTextWidth hangs
   /** get the text width in pixels */
   // @ts-ignore
+  await timeout(async () => {
+    try { dojo.require('dojo.ready'); } catch (err: any) {
+      _package.logger.warning(err.message);
+    }
+    try { dojo.require('dojo.window'); } catch (err: any) {
+      _package.logger.warning(err.message);
+    }
+    try { dojo.require('dojo.io.script'); } catch (err: any) {
+      _package.logger.warning(err.message);
+    }
+    try { dojo.require('dojo.request.iframe'); } catch (err: any) {
+      _package.logger.warning(err.message);
+    }
+    try { dojo.require('dojo.dom'); } catch (err: any) {
+      _package.logger.warning(err.message);
+    }
+    try { dojo.require('dojox.gfx'); } catch (err: any) {
+      _package.logger.warning(err.message);
+    }
+    try { dojo.require('dojox.gfx.svg'); } catch (err: any) {
+      _package.logger.warning(err.message);
+    }
+    try { dojo.require('dojox.gfx.shape'); } catch (err: any) {
+      _package.logger.warning(err.message);
+    }
+    // @formatter:off
+    while (!(
+      dojo && dojo.window && dojo.io?.script && dojo.request?.iframe &&
+      dojox && dojox.gfx && dojox.gfx?.svg && dojox.gfx?.shape && !dojox.gfx?.createSurface
+    )) await delay(100);
+    // @formatter:on
+  }, 60000, 'timeout dojox.gfx.svg');
+
+  _package.logger.debug(`${logPrefix}, patch window.dojox.gfx.svg.Text.prototype.getTextWidth`);
+  // @ts-ignore
   window.dojox.gfx.svg.Text.prototype.getTextWidth = function() {
+    // Patched via Datagrok Helm package
     const rawNode = this.rawNode;
     const oldParent = rawNode.parentNode;
     const _measurementNode = rawNode.cloneNode(true);
@@ -57,6 +102,7 @@ export function initHelmPatchDojo(): void {
     oldParent.removeChild(_measurementNode);
     return _width;
   };
+  _package.logger.debug(`${logPrefix}, end`);
 }
 
 export const helmJsonReplacer = (key: string, value: any): any => {
@@ -70,47 +116,34 @@ export const helmJsonReplacer = (key: string, value: any): any => {
 };
 
 export class HelmPackage extends DG.Package {
-  public getMonomerOriginal: GetMonomerFunc;
   public alertOriginal: (s: string) => void;
+  public readonly hh: HelmHelper;
+
+  constructor() {
+    super();
+    this.hh = new HelmHelper(this.logger);
+  }
+
+  public initHelmPatchScilAlert(): void {
+    const logPrefix: string = 'Helm: initHelmPatchScilAlert()';
+    this.alertOriginal = scil.Utils.alert;
+    scil.Utils.alert = (s: string): void => {
+      this.logger.warning(`${logPrefix}, scil.Utils.alert() s: "${s}".`);
+    };
+  }
 
   /** Patches Pistoia Monomers.getMonomer method to utilize DG Bio monomer Lib */
-  public initHelmPatchPistoia(monomerLib: IMonomerLib, logger: ILogger): void {
+  public async initHelmPatchPistoia(monomerLib: IMonomerLib): Promise<void> {
     const logPrefix: string = 'Helm: initHelmPatchPistoia()';
     const monomers = org.helm.webeditor.Monomers;
 
-    this.getMonomerOriginal = monomers.getMonomer.bind(monomers);
     this.logger.debug(`${logPrefix}, this.getMonomerOriginal stored`);
 
-    this.alertOriginal = scil.Utils.alert;
-
-    org.helm.webeditor.Monomers.getMonomer = (
-      a: Atom<HelmType> | HelmType, name: string
-    ): GetMonomerResType => {
-      const logPrefixInt = `${logPrefix}, org.helm.webeditor.Monomers.getMonomer()`;
-      try {
-        // logger.debug(`${logPrefixInt}, a: ${JSON.stringify(a, helmJsonReplacer)}, name: '${name}'`);
-
-        // Creates monomers in lib
-        const dgWem = getWebEditorMonomer(monomerLib, a, name);
-
-        // // Returns null for gap
-        // const oWem = this.getMonomerOriginal(a, name);
-        // if (!oWem)
-        //   logger.warning(`${logPrefixInt}, getMonomerOriginal( a: ${a}, name: ${name}) returns null`);
-        return dgWem; //dgWem;
-      } catch (err) {
-        const [errMsg, errStack] = errInfo(err);
-        logger.error(`${logPrefixInt}, Error: ${errMsg}`, undefined, errStack);
-        throw err;
-      }
-    };
-    scil.Utils.alert = (s: string): void => {
-      logger.warning(`${logPrefix}, scil.Utils.alert() s = 's'.`);
-    };
+    this.hh.overrideGetMonomer(this.hh.buildGetMonomerFromLib(monomerLib));
 
     // @ts-ignore, intercept with proxy to observe access and usage
     org.helm.webeditor.Monomers = new class {
-      constructor(base: org.helm.IMonomers) {
+      constructor(base: Monomers) {
         return new Proxy(base, {
           get(target: any, p: string | symbol, _receiver: any): any {
             return target[p];
@@ -127,11 +160,11 @@ export class HelmPackage extends DG.Package {
           },
         });
       }
-    }(org.helm.webeditor.Monomers) as org.helm.IMonomers;
+    }(org.helm.webeditor.Monomers) as Monomers;
 
     // @ts-ignore, intercept with proxy to observe access and usage
     org.helm.webeditor = new class {
-      constructor(base: org.helm.IOrgHelmWebEditor) {
+      constructor(base: IOrgHelmWebEditor) {
         return new Proxy(base, {
           get(target: any, p: string | symbol, _receiver: any): any {
             return target[p];
@@ -145,6 +178,6 @@ export class HelmPackage extends DG.Package {
           },
         });
       }
-    }(org.helm.webeditor) as org.helm.IOrgHelmWebEditor;
+    }(org.helm.webeditor) as IOrgHelmWebEditor;
   }
 }